backend/requirements.txt
Flask==3.0.1
Flask-Cors==4.0.0
openai==1.9.0
python-dotenv==1.0.0
pymongo==4.6.1
pyjwt==2.8.0
passlib==1.7.4
asgiref==3.7.2